+++ src/main/java/nextstep/security/context/DelegatingSecurityContextRepository.java
@@ -2,8 +2,6 @@ package nextstep.security.context;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.security.SecurityContext;
-import nextstep.security.SecurityContextRepository;
 
 import java.util.Arrays;
 import java.util.List;
