+++ src/main/java/nextstep/security/context/SecurityContext.java
@@ -1,4 +1,6 @@
-package nextstep.security;
+package nextstep.security.context;
+
+import nextstep.security.authentication.Authentication;
 
 public interface SecurityContext {
     Authentication getAuthentication();
