+++ src/main/java/nextstep/security/converter/UsernamePasswordAuthenticationConverter.java
@@ -1,11 +1,8 @@
 package nextstep.security.converter;
 
 import jakarta.servlet.http.HttpServletRequest;
-import nextstep.security.Authentication;
-import nextstep.security.AuthenticationConverter;
+import nextstep.security.authentication.Authentication;
 import nextstep.security.authentication.UsernamePasswordAuthenticationToken;
-
-import java.util.Map;
 
 public class UsernamePasswordAuthenticationConverter implements AuthenticationConverter {
