+++ src/main/java/nextstep/security/filter/BasicAuthenticationFilter.java
@@ -4,11 +4,11 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.security.Authentication;
-import nextstep.security.AuthenticationConverter;
-import nextstep.security.AuthenticationManager;
-import nextstep.security.SecurityContext;
-import nextstep.security.SecurityContextRepository;
+import nextstep.security.authentication.Authentication;
+import nextstep.security.converter.AuthenticationConverter;
+import nextstep.security.authentication.AuthenticationManager;
+import nextstep.security.context.SecurityContext;
+import nextstep.security.context.SecurityContextRepository;
 import nextstep.security.context.SecurityContextHolder;
 import nextstep.security.exception.AuthenticationException;
 import org.slf4j.Logger;
