+++ src/main/java/nextstep/security/filter/FilterChainProxy.java
@@ -6,7 +6,6 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
-import nextstep.security.SecurityFilterChain;
 import org.springframework.web.filter.GenericFilterBean;
 
 import java.io.IOException;
