+++ src/main/java/nextstep/security/filter/SecurityContextHolderFilter.java
@@ -6,8 +6,8 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.security.SecurityContext;
-import nextstep.security.SecurityContextRepository;
+import nextstep.security.context.SecurityContext;
+import nextstep.security.context.SecurityContextRepository;
 import nextstep.security.context.SecurityContextHolder;
 import org.springframework.web.filter.GenericFilterBean;
