+++ src/main/java/nextstep/security/userdetails/UserDetailsService.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.userdetails;
 
 public interface UserDetailsService {
     UserDetails loadUserByUsername(String username);
